Overhaul the coding style, fix various bugs, and move towards getting the interpreter working.
Remove some more obsolete files.

+++ src/hash.h
@@ -23,16 +23,16 @@ typedef struct {
 } HashTable;
 
 typedef struct {
-  uintptr_t entry;
+  uint32_t entry;
   HashNode *prev;
   HashNode *curr;
   int found;
 } HashLink;
 
-HashTable *hashNew(uintptr_t size);
-void hashDestroy(HashTable *table);
-void hashGet(HashTable *table, char *key, HashLink *l);
-HashNode *hashSet(HashTable *table, HashLink *l, char *key, void *body);
-uintptr_t strHash(char *str);
+HashTable *hashNew (uintptr_t size);
+void hashDestroy (HashTable *table);
+void hashGet (HashTable *table, char *key, HashLink *l);
+HashNode *hashSet (HashTable *table, HashLink *l, char *key, void *body);
+uint32_t strHash (char *str);
 
 #endif
